fix(config) ignored with the option unknown-config-key
Currently, the unknown_config_key option ensures that keys in the config
ini file that are not known are just ignored and do not raise an error.

Similarly it is possible that there are cmd-arguments which the current
Pydantic Model does not understand. In the current implementation,
additional cmd-line-arguments that cannot be parsed raise an error.

To properly align with the behaviour of unknown-config-key in ini
files, this little modification ensures that undefined
command-line-arguments are just ignored instead of raising an error.

If unknown_config_key is not set to "ignore", the behaviour is the same
as before and an error is raised.